Playing with Crafty and Coffeescript.
Added simplistic resource loading. In import_modules, you can define
code modules to load with the IMPORT_MODULES object.

You can also define images to preload in the GRAPHICS_MAPPING object. If
the image also has an object attached, this is treated as a definition
for creating a sprite using Craft.sprite.